precise GC: STORE 漏れ compile-time audit + 64GB round-robin PURGE + scramble 廃止
framework / sample 両側で precise GC の audit 機構を再整理。 2 phase 分の作業を
1 commit にまとめる。

## Phase 2a: STORE 漏れ compile-time 検出

- gc_types.h: ARO_GC_EDGE qualifier 追加 (= ARO_GC_WB_AUDIT=1 build で const)
- gc.h: ARO_STORE/_BULK macro が (VALUE *)(uintptr_t)(slot) で const 抜き、
        sample の事故 cast を Werror で reject
- Makefile (baruby/ascheme): ARO_GC_WB_AUDIT=1 で
  -DARO_GC_WB_AUDIT -Werror=discarded-qualifiers -Werror=cast-qual の 3 点セット
- sample heap ptr field (BaArray.items / pair.car / sframe.slots 等) を qualify
- runtime/astro_build.c + sample/baruby_precise/astro_jit.c の
  free((void *)const_*) を (void *)(uintptr_t) 経由に統一

audit で catch される 3 パターン:
  a->items = x;                          → error: assignment of read-only member
  memcpy(&a->items, ...);                → -Werror=discarded-qualifiers
  memcpy((void *)&a->items, ...);        → -Werror=cast-qual

## Phase 3: scramble 廃止 + 64GB round-robin PURGE

- runtime/precise_gc/gc_copy_scramble.c 削除 (= backend ID 17、 永久欠番)
- ARO_GC_HAS_SCRAMBLE flag、 scramble_R / scramble_R_old field 削除
- ARO_TOUCH / ARO_ENCODE macro 削除、 ARO_LOAD は plain *slot に simplify
- ARO_GC_VISIT_EDGE から XOR encode/decode 削除
- gc_copy.c: BARUBY_GC_PURGE=1 を 64 GiB virtual round-robin mprotect に refactor
    - init で 64 GiB MAP_NORESERVE PROT_NONE reserve
    - 各 GC で次 plane (= 64 MiB) を mprotect R/W、 旧 from-space を
      mprotect PROT_NONE + madvise DONTNEED
    - 1024 plane 巡回、 wrap 時に R/W 再有効化
- sample API 簡素化: VAL2ARY(c, v) / VAL2STR(c, v) を直接 cast に
- baruby_precise の manual XOR 2 site も削除

新 audit catch 範囲 (= 旧 scramble + R rotation の superset):
- visit_edge 漏れ: 次 GC scan 時の forward read で確定 SEGV
- decoded raw in C local across GC: deref で from-space PROT_NONE 領域 → SEGV
- 「たまたま動く」 確率を 1024-plane lag で実質ゼロ化

## docs

- docs/gc_design.md §3.3: audit 機構 (Layer 1 compile-time + Layer 2 runtime PURGE) に書直し
- docs/gc_design.md §6: Phase 2a + Phase 3 を追記
- docs/gc_design.md §7.7: 検証戦略を audit + stress + purge に更新
- docs/wb_audit.md: 実装済 status 追加 (= 元設計メモは保存)
- docs/gc_runtime.md: copy_scramble entry を round-robin PURGE 言及に置換

## 検証

- baruby_precise: 15 backend × default/audit smoke 全 pass
- ascheme_precise: 同上 + oracle 18/18 + R5RS 179/179 + test_stress 18/18 pass
- audit 違反 3 パターン全 compile error 化を実証

## 既知 pre-existing flake

- baruby_precise/test_p1.ba.rb は STRESS+PURGE で SEGV (HEAD でも再現、 別途調査)

32 files changed, 599 insertions(+), 1089 deletions(-)
net -490 行 (= scramble feature 削除 + API 簡素化)

+++ runtime/astro_build.c
@@ -253,22 +253,22 @@ astro_build_config_dispose(struct astro_build_config *cfg)
 {
     // Free heap-allocated strings (cc / sanitize from env parse) and
     // the heap arrays for cflags / ldflags.
-    free((void *)cfg->cc);
+    free((void *)(uintptr_t)cfg->cc);
     cfg->cc = NULL;
-    free((void *)cfg->sanitize);
+    free((void *)(uintptr_t)cfg->sanitize);
     cfg->sanitize = NULL;
     if (cfg->extra_cflags) {
         for (const char *const *p = cfg->extra_cflags; *p; p++) {
-            free((void *)*p);
+            free((void *)(uintptr_t)*p);
         }
-        free((void *)cfg->extra_cflags);
+        free((void *)(uintptr_t)cfg->extra_cflags);
         cfg->extra_cflags = NULL;
     }
     if (cfg->extra_ldflags) {
         for (const char *const *p = cfg->extra_ldflags; *p; p++) {
-            free((void *)*p);
+            free((void *)(uintptr_t)*p);
         }
-        free((void *)cfg->extra_ldflags);
+        free((void *)(uintptr_t)cfg->extra_ldflags);
         cfg->extra_ldflags = NULL;
     }
 }
@@ -500,7 +500,7 @@ astro_build_executable(const struct astro_build_config *cfg)
             fprintf(stderr, "astro_build: strip failed (exit %d)\n", sret);
         }
         free(strip_cmd);
-        if (strip_prog != (const char *)"strip") free((void *)strip_prog);
+        if (strip_prog != (const char *)"strip") free((void *)(uintptr_t)strip_prog);
     }
 
     return ret;
@@ -594,7 +594,7 @@ astro_build_aot_executable(struct Node *root,
         unlink(embed_path);
     }
     cfg->extra_sources_abs = prev_extras;
-    for (uint32_t i = 0; i < n_sd; i++) free((void *)sd_paths[i]);
+    for (uint32_t i = 0; i < n_sd; i++) free((void *)(uintptr_t)sd_paths[i]);
     free(sd_paths);
     free(extras);
     return rc;
